requirements.txt
python-igraph
numpy
matplotlib
tqdm
ortools
pycairo
pandas
scipy
gurobipy